requirements.txt
Flask==2.2.2
spotipy==2.22.0
pandas
gunicorn==20.1.0
opencv_python_headless==4.7.0.68
numpy
Pillow==9.4.0
tensorflow==2.11.0
